Numeric.LinearAlgebra.HMatrix

Description

compatibility with previous version, to be removed

Documentation

module Numeric.LinearAlgebra

(¦) :: Matrix Double -> Matrix Double -> Matrix Double infixl 3

a synonym for (|||) (unicode 0x00a6, broken bar)

—— :: Matrix Double -> Matrix Double -> Matrix Double infixl 2

a synonym for (===) (unicode 0x2014, em dash)

type ℝ = Double

type ℂ = Complex Double

<·> :: Numeric t => Vector t -> Vector t -> t infixr 8

app :: Numeric t => Matrix t -> Vector t -> Vector t

mul :: Numeric t => Matrix t -> Matrix t -> Matrix t

cholSH :: Field t => Matrix t -> Matrix t

Similar to chol, without checking that the input matrix is hermitian or symmetric. It works with the upper triangular part.

mbCholSH :: Field t => Matrix t -> Maybe (Matrix t)

Similar to cholSH, but instead of an error (e.g., caused by a matrix not positive definite) it returns Nothing.

eigSH' :: Field t => Matrix t -> (Vector Double, Matrix t)

Similar to eigSH without checking that the input matrix is hermitian or symmetric. It works with the upper triangular part.

eigenvaluesSH' :: Field t => Matrix t -> Vector Double

Similar to eigenvaluesSH without checking that the input matrix is hermitian or symmetric. It works with the upper triangular part.

geigSH'

Arguments

:: Field t

=> Matrix t
A

-> Matrix t
B

-> (Vector Double, Matrix t)