Numeric.LinearAlgebra.Static

Description

Experimental interface with statically checked dimensions.

See code examples at http://dis.um.es/~alberto/hmatrix/static.html.

Vector

type ℝ = Double

data R n

vec2 :: ℝ -> ℝ -> R 2

vec3 :: ℝ -> ℝ -> ℝ -> R 3

vec4 :: ℝ -> ℝ -> ℝ -> ℝ -> R 4

(&) :: forall n. (KnownNat n, 1 <= n) => R n -> ℝ -> R (n + 1) infixl 4

(#) :: forall n m. (KnownNat n, KnownNat m) => R n -> R m -> R (n + m) infixl 4

split :: forall p n. (KnownNat p, KnownNat n, p <= n) => R n -> (R p, R (n - p))

headTail :: (KnownNat n, 1 <= n) => R n -> (ℝ, R (n - 1))

vector :: KnownNat n => [ℝ] -> R n

linspace :: forall n. KnownNat n => (ℝ, ℝ) -> R n

range :: forall n. KnownNat n => R n

dim :: forall n. KnownNat n => R n

Matrix

data L m n

type Sq n = L n n

build :: forall m n. (KnownNat n, KnownNat m) => (ℝ -> ℝ -> ℝ) -> L m n

row :: R n -> L 1 n

col :: KnownNat n => R n -> L n 1

(|||) :: (KnownNat r1, KnownNat c, KnownNat r2, KnownNat ((+) r1 r2)) => L c r1 -> L c r2 -> L c ((+) r1 r2) infixl 3

(===) :: (KnownNat r1, KnownNat r2, KnownNat c) => L r1 c -> L r2 c -> L (r1 + r2) c infixl 2

splitRows :: forall p m n. (KnownNat p, KnownNat m, KnownNat n, p <= m) => L m n -> (L p n, L (m - p) n)

splitCols :: forall p m n. (KnownNat p, KnownNat m, KnownNat n, KnownNat (n - p), p <= n) => L m n -> (L m p, L m (n - p))

unrow :: L 1 n -> R n

uncol :: KnownNat n => L n 1 -> R n

tr :: Transposable m mt => m -> mt

conjugate transpose

eye :: KnownNat n => Sq n

diag :: KnownNat n => R n -> Sq n

blockAt :: forall m n. (KnownNat m, KnownNat n) => ℝ -> Int -> Int -> Matrix Double -> L m n

matrix :: (KnownNat m, KnownNat n) => [ℝ] -> L m n

Complex

data C n

data M m n

data Her n

her :: KnownNat n => M n n -> Her n

𝑖 :: Sized ℂ s c => s

Products

(<>) :: forall m k n. (KnownNat m, KnownNat k, KnownNat n) => L m k -> L k n -> L m n infixr 8

(#>) :: (KnownNat m, KnownNat n) => L m n -> R n -> R m infixr 8

(<.>) :: R n -> R n -> ℝ infixr 8

Linear Systems

linSolve :: (KnownNat m, KnownNat n) => L m m -> L m n -> Maybe (L m n)

(<\>) :: (KnownNat m, KnownNat n, KnownNat r) => L m n -> L m r -> L n r

Factorizations

svd :: (KnownNat m, KnownNat n) => L m n -> (L m m, R n, L n n)

withCompactSVD :: forall m n z. (KnownNat m, KnownNat n) => L m n -> (forall k. KnownNat k => (L m k, R k, L n k) -> z) -> z

svdTall :: (KnownNat m, KnownNat n, n <= m) => L m n -> (L m n, R n, L n n)

svdFlat :: (KnownNat m, KnownNat n, m <= n) => L m n -> (L m m, R m, L n m)

class Eigen m l v | m -> l, m -> v where

Methods

eigensystem :: m -> (l, v)

eigenvalues :: m -> l

withNullspace :: forall m n z. (KnownNat m, KnownNat n) => L m n -> (forall k. KnownNat k => L n k -> z) -> z

withOrth :: forall m n z. (KnownNat m, KnownNat n) => L m n -> (forall k. KnownNat k => L n k -> z) -> z

qr :: (KnownNat m, KnownNat n) => L m n -> (L m m, L m n)

chol :: KnownNat n => Sym n -> Sq n

Norms

class Normed a where

Methods

norm_0 :: a -> R

norm_1 :: a -> R

norm_2 :: a -> R

norm_Inf :: a -> R

Random arrays

type Seed = Int

data RandDist

Constructors

Uniform
uniform distribution in [0,1)

Gaussian
normal distribution with mean zero and standard deviation one

randomVector :: forall n. KnownNat n => Seed -> RandDist -> R n

rand :: forall m n. (KnownNat m, KnownNat n) => IO (L m n)

randn :: forall m n. (KnownNat m, KnownNat n) => IO (L m n)

gaussianSample :: forall m n. (KnownNat m, KnownNat n) => Seed -> R n -> Sym n -> L m n

uniformSample

Arguments

:: (KnownNat m, KnownNat n)

=> Seed

-> R n
minimums of each row

-> R n
maximums of each row

-> L m n

Misc

mean :: (KnownNat n, 1 <= n) => R n -> ℝ

meanCov :: forall m n. (KnownNat m, KnownNat n) => L m n -> (R n, Sym n)

class Disp t where

Methods

disp :: Int -> t -> IO ()

class Domain field vec mat | mat -> vec field, vec -> mat field, field -> mat vec where

Methods

mul :: forall m k n. (KnownNat m, KnownNat k, KnownNat n) => mat m k -> mat k n -> mat m n

app :: forall m n. (KnownNat m, KnownNat n) => mat m n -> vec n -> vec m

dot :: forall n. KnownNat n => vec n -> vec n -> field

cross :: vec 3 -> vec 3 -> vec 3

diagR :: forall m n k. (KnownNat m, KnownNat n, KnownNat k) => field -> vec k -> mat m n

dvmap :: forall n. KnownNat n => (field -> field) -> vec n -> vec n

dmmap :: forall n m. (KnownNat m, KnownNat n) => (field -> field) -> mat n m -> mat n m

outer :: forall n m. (KnownNat m, KnownNat n) => vec n -> vec m -> mat n m

zipWithVector :: forall n. KnownNat n => (field -> field -> field) -> vec n -> vec n -> vec n

withVector :: forall z. Vector ℝ -> (forall n. KnownNat n => R n -> z) -> z

withMatrix :: forall z. Matrix ℝ -> (forall m n. (KnownNat m, KnownNat n) => L m n -> z) -> z

exactLength :: forall n m. (KnownNat n, KnownNat m) => R m -> Maybe (R n)

Useful for constraining two dependently typed vectors to match each other in length when they are unknown at compile-time.

exactDims :: forall n m j k. (KnownNat n, KnownNat m, KnownNat j, KnownNat k) => L m n -> Maybe (L j k)

Useful for constraining two dependently typed matrices to match each other in dimensions when they are unknown at compile-time.

toRows :: forall m n. (KnownNat m, KnownNat n) => L m n -> [R n]

toColumns :: forall m n. (KnownNat m, KnownNat n) => L m n -> [R m]

withRows :: forall n z. KnownNat n => [R n] -> (forall m. KnownNat m => L m n -> z) -> z

withColumns :: forall m z. KnownNat m => [R m] -> (forall n. KnownNat n => L m n -> z) -> z

class Num t => Sized t s d | s -> t, s -> d where

Methods

konst :: t -> s

unwrap :: s -> d t

fromList :: [t] -> s

extract :: s -> d t

create :: d t -> Maybe s

size :: s -> IndexOf d

class Diag m d | m -> d where

Methods

takeDiag :: m -> d

data Sym n

sym :: KnownNat n => Sq n -> Sym n

mTm :: (KnownNat m, KnownNat n) => L m n -> Sym n

unSym :: Sym n -> Sq n

<·> :: R n -> R n -> ℝ infixr 8